Description: fix FUZZ error.
Match-id-17710166afb957db77a640306215d466d1aa4dd1

+++ hdf_service/distributed_audio/hdi_service/audio/v1_0/test/fuzztest/captureframe_fuzzer/captureframe_fuzzer.cpp
@@ -29,7 +29,7 @@ namespace Audio {
 namespace V1_0 {
 void CaptureFrameFuzzTest(const uint8_t* data, size_t size)
 {
-    if ((data == nullptr) || (size < (sizeof(int32_t)))) {
+    if ((data == nullptr) || (size < (sizeof(uint64_t)))) {
         return;
     }
